feat: Refactor ChatInterface with hooks and state management
- Replace legacy ChatInterface with a memoized functional component using
  custom hooks for workflow execution, collection, scroll, and chat state
  management.
- Remove redundant state variables and handlers; centralize state via useChatState.
- Modularize logic: separate concerns for workflow execution, input handling,
  and UI state toggling.
- Introduce ChatContainer component to encapsulate chat area and PDF viewer,
  improving layout and performance.
- Simplify additionalParams validation with memoized function.
- Improve scroll handling with useScrollManagement hook for better UX.
- Remove deprecated dynamic imports and ResizablePanel in favor of integrated
  panel management.
- Clean up event handlers and side
