+++ backup/ChatInterfaceRefactored2.tsx
