+++ tsconfig.json
@@ -29,5 +29,8 @@
         ".next/types/**/*.ts",
         "src"
     ],
-    "exclude": ["node_modules"]
+    "exclude": [
+        "node_modules",
+        "backup"
+    ]
 }
